Move utility functions out of datamodels.schema
